Fixes to build process and to ChromeBrowser for current Windows versions of Chrome.
Updated Maven "fat" target to "jar-with-dependencies".
Updated Maven source/target JDK version to 1.9.
Updated ChromeBrowser to look for files ending in "Cookies" in any subdirectory, to handle modern Windows Chrome that puts it in Default/Network.
Updated ChromeBrowser to skip un-decryptable cookies in Windows, rather than fail with an exception, to match Mac and Linux behavior.